% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidModels-package.R
\docType{package}
\name{covidModels}
\alias{covidModels}
\alias{covidModels-package}
\title{covidModels - Predictive models for COVID-19}
\description{
Predictive models for COVID-19
}
\author{
Evan L. Ray \href{mailto:eray@mtholyoke.edu}{eray@mtholyoke.edu}
}
