% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_baseline.R
\name{fit_quantile_baseline}
\alias{fit_quantile_baseline}
\title{Fit a quantile baseline model to historical disease incidence}
\usage{
fit_quantile_baseline(incidence)
}
\arguments{
\item{incidence}{numeric vector of disease incidence in past time points}
}
\value{
quantile_baseline fit object
}
\description{
Fit a quantile baseline model to historical disease incidence
}
