% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_smoother.R
\name{fit_spline_smoother}
\alias{fit_spline_smoother}
\title{Fit a quantile baseline model to historical disease incidence}
\usage{
fit_spline_smoother(
  incidence,
  temporal_resolution = "weekly",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{incidence}{numeric vector of disease incidence in past time points}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: 'weekly' or 'daily'.  If 'daily', a day-of-week offset is included.}

\item{verbose}{logical: if TRUE, print output from estimation in Stan}

\item{...}{other arguments are ignored}
}
\value{
spline_smoother fit object
}
\description{
Fit a quantile baseline model to historical disease incidence
}
