% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_trend.R
\name{is.local_trend}
\alias{is.local_trend}
\title{Check if object is of class local_trend}
\usage{
is.local_trend(object)
}
\arguments{
\item{object}{an object that may be a local_trend object}
}
\value{
boolean; whether object is inherits local_trend class
}
\description{
Check if object is of class local_trend
}
