% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_smoother.R
\name{is.spline_smoother}
\alias{is.spline_smoother}
\title{Check if object is of class spline_smoother}
\usage{
is.spline_smoother(object)
}
\arguments{
\item{object}{an object that may be a spline_smoother object}
}
\value{
boolean; whether object is inherits spline_smoother class
}
\description{
Check if object is of class spline_smoother
}
