% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_trend.R
\name{new_local_trend}
\alias{new_local_trend}
\title{Create a local_trend object}
\usage{
new_local_trend(inc_diffs)
}
\arguments{
\item{inc_diffs}{historical first differences in incidence}
}
\value{
local_trend fit object
}
\description{
Create a local_trend object
}
