% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_baseline.R
\name{new_predict.quantile_baseline}
\alias{new_predict.quantile_baseline}
\title{Predict future disease incidence by resampling one-step-ahead forecasts}
\usage{
new_predict.quantile_baseline(quantile_baseline, newdata, horizon, nsim, ...)
}
\arguments{
\item{quantile_baseline}{a quantile_baseline fit object}

\item{newdata}{numeric vector of length at least one with incident counts}

\item{horizon}{number of time steps forward to predict}

\item{nsim}{number of samples to use for generating predictions at
horizons greater than 1}

\item{...}{mop up unused arguments}
}
\value{
matrix of simulated incidence with nsim rows and horizon columns
}
\description{
Predict future disease incidence by resampling one-step-ahead forecasts
}
