% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_baseline.R
\name{new_quantile_baseline}
\alias{new_quantile_baseline}
\title{Create a quantile_baseline object}
\usage{
new_quantile_baseline(inc_diffs)
}
\arguments{
\item{inc_diffs}{historical first differences in incidence}
}
\value{
quantile_baseline fit object
}
\description{
Create a quantile_baseline object
}
