% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_smoother.R
\name{new_spline_smoother}
\alias{new_spline_smoother}
\title{Create a spline_smoother object}
\usage{
new_spline_smoother(
  param_estimates,
  temporal_resolution,
  train_incidence,
  spline_order,
  interior_knots,
  boundary_knots
)
}
\arguments{
\item{param_estimates}{model parameter estimates}

\item{temporal_resolution}{character vector specifying temporal resolution
to include: 'weekly' or 'daily'.  If 'daily', a day-of-week offset is included.}

\item{train_incidence}{integer vector of training set incidence}

\item{spline_order}{order of the spline for the mean}

\item{interior_knots}{locations of interior knots}

\item{bounary_knots}{locations of boundary knots}
}
\value{
spline_smoother fit object
}
\description{
Create a spline_smoother object
}
