% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_baseline.R
\name{predict.quantile_baseline}
\alias{predict.quantile_baseline}
\title{Predict future disease incidence by resampling one-step-ahead forecasts}
\usage{
\method{predict}{quantile_baseline}(quantile_baseline, inc_data, cum_data, quantiles, horizon, num_samples)
}
\arguments{
\item{quantile_baseline}{a quantile_baseline fit object}

\item{inc_data}{numeric vector of length at least one with incident counts}

\item{cum_data}{numeric vector of length at least one with cumulative counts}

\item{quantiles}{quantile levels for which  to generate predictions}

\item{horizon}{number of time steps forward to predict}

\item{num_samples}{number of samples to use for generating predictions at
horizons greater than 1}
}
\value{
data frame with columns target, quantile, and value with forecasts
of incident and cumulative deaths
}
\description{
Predict future disease incidence by resampling one-step-ahead forecasts
}
