% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_smoother.R
\name{predict.spline_smoother}
\alias{predict.spline_smoother}
\title{Predict future disease incidence starting from the end of the training data.}
\usage{
\method{predict}{spline_smoother}(
  spline_smoother,
  newdata,
  temporal_resolution,
  post_pred,
  horizon,
  nsim
)
}
\arguments{
\item{spline_smoother}{a spline_smoother fit object}

\item{temporal_resolution}{character vector specifying temporal resolution
to predict at: 'weekly' or 'daily'. The 'daily' option is only valid if the
model was fit at a daily resolution.}

\item{post_pred}{logical; if TRUE, samples from the posterior predictive are
returned for time points with observations in the training set}

\item{horizon}{number of time steps forward to predict}

\item{nsim}{number of samples to use for generating predictions at
horizons greater than 1}

\item{...}{other arguments are ignored}
}
\value{
matrix of simulated incidence with nsim rows and horizon columns
}
\description{
Predict future disease incidence starting from the end of the training data.
}
